Improve read_labels() to notify invalid input
Improve the read_labels() to return a triplet consisting of the label, an
error flag and an (optional) error message. This allows to prevently
mark some labels as invalid in the LGR and provide the user with some
explanations.
As a consequence of the clean-up, remove now unused read_set_labels().