Fix cross-script variant tool
Update the tool to process an LGR (can be a regular LGR or a set), and a
label file. When one of the variant of a label present in the input file
contains one (or more) code points that do not have the same script as
the LGR, then display these CPs.