Improve speed of LGR merge
Improve the merge_chars function by a factor ~45.
Profiling has shown that an abnormal amount of time (both in sec and
number of calls) was spent in Char.__eq__. Replace iteration in set of
merged characters by dict indexed with has, as it will be more efficient
to find a specific object.