Fix generation of prefix list with a code point sequence a look-around rule
In some specific cases validation of a label's variants was failing
because of an out-of-bound error.
By using a variant producing a code point sequence, we can end up in the
case where prefixed_label is larger than orig_label.
If the current "prefix" char has a contextual rule that invovolves
look-around rules, then this caused the regular expression to match on
orig_label with a start index larger than the original label.
Fix this issue by creating the prefixed label as it is done for variant
labels: take the current prefix, add the considered character, and
append the remaining of the label.