Make collision tools do not stop on not-in-LGR labels
Until now, the collision tools were stopping (raising a NotInLgr
exception) when one of the input label was not eligible (~not in the
LGR). This was bothersome for users as it is expected in some cases to
have out of LGR labels in the input file (eg. to check any collisions
created by cross-script variants in the same LGR).
This fix allows the tool to list the invalid labels at the top of the
output, and continue processing only valid labels.