Store rule name(s) when code point does not comply with a contextual rule
When validating a label, store the list of rule names defined for code
points that do not validate. As a code point can be used as the start of
multiple <char> elements in an LGR, we will only return a "flat" list
associated with the first code point starting with the input code points
value.